refactor CompilerMatcher.scala
